+++ noodles-cram/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.68.0 - 2024-09-04
 
 ### Added